cheri: Fix typecasts in fs/ext4
- Reference counts are actually integers stored in a pointer.
- 32-bit cast missed by automatic conversions
- Fake a pointer from an instruction pointer address in tracing.

+++ fs/ext4/ioctl.c
@@ -1535,7 +1535,7 @@ static long __ext4_ioctl(struct file *filp, unsigned int cmd, user_uintptr_t arg
 	struct super_block *sb = inode->i_sb;
 	struct mnt_idmap *idmap = file_mnt_idmap(filp);
 
-	ext4_debug("cmd = %u, arg = %lu\n", cmd, arg);
+	ext4_debug("cmd = %u, arg = %lu\n", cmd, (unsigned long)arg);
 
 	switch (cmd) {
 	case FS_IOC_GETFSMAP:
